requirements.txt
flask
groq
python-dotenv
easyocr
pymupdf
pillow
